dev(justfile): added `compose-up` and `migrate` rules

+++ justfile
@@ -9,7 +9,14 @@ run:
 
 start: build run
 
+compose-up:
+	docker-compose up -d
+
+migrate: erase-db-data compose-up
+	sleep 5
+	go run ./cmd/migrator/
+
 erase-db-data:
 	docker kill $(docker ps -qa) 2> /dev/null || true
 	docker rm -f $(docker ps -qa) 2> /dev/null || true
-	sudo rm -rf ./_data/
+	if test -d "./_data"; then sudo rm -rf ./_data/; fi
